requirements.txt
imgaug==0.4.0
keras==3.3.3
matplotlib==3.8.3
numpy==1.26.4
pandas==2.2.2
Pillow==9.0.1
Pillow==10.3.0
protobuf==3.12.4
scikit_learn==1.4.1.post1
scipy==1.13.1
seaborn==0.13.2
skimage==0.0
tensorflow==2.16.1
tqdm==4.66.4
